# api.yaml
openapi: 3.0.0
info:
  title: MediCore API
  description: Healthcare management system API for patients, doctors, and administrators
  version: 1.0.0
  contact:
    name: MediCore Support
servers:
  - url: http://localhost:5000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BookingRequest:
      type: object
      required:
        - date
        - time
      properties:
        date:
          type: string
          format: date
        time:
          type: string
          enum: ['09:00 AM','10:00 AM','11:00 AM','12:00 PM','02:00 PM','03:00 PM','04:00 PM','05:00 PM']
        reason:
          type: string
    PaymentRequest:
      type: object
      required:
        - payment_method
        - amount
      properties:
        payment_method:
          type: string
          enum: [card, upi]
        amount:
          type: number
        booking_id:
          type: integer
        card_number:
          type: string
        expiry:
          type: string
        cvv:
          type: string
        upi_id:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /patient/profile:
    get:
      tags:
        - Patient
      summary: Get patient profile
      responses:
        '200':
          description: Patient profile
    put:
      tags:
        - Patient
      summary: Update patient profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                gender:
                  type: string
                blood_type:
                  type: string
                location:
                  type: string
                birth_date:
                  type: string
                  format: date
      responses:
        '200':
          description: Profile updated
  /patient/physicians:
    get:
      tags:
        - Patient
      summary: Search physicians
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query
      responses:
        '200':
          description: List of physicians
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /patient/bookings/{pid}:
    post:
      tags:
        - Patient
      summary: Book appointment with physician
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '201':
          description: Booking created
        '400':
          description: Invalid request
        '409':
          description: Time slot already booked
  /patient/bookings/{id}/cancel:
    put:
      tags:
        - Patient
      summary: Cancel booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking cancelled
        '403':
          description: Unauthorized
...
  /patient/payments:
    post:
      tags:
        - Patient
      summary: Process payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment processed
        '400':
          description: Invalid payment details
...
  /patient/task/{task_id}:
    get:
      tags:
        - Patient
      summary: Check export task status
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task status
...